Added beginning of a light node class.
Class that I want to use for the scene graph that will mark where lights
are. When maybe a debug flag is set they will render as some spheres or
something... Not sure how that will work with object nodes that shader's
code rely on them, but I have the same problem with Cameras.